pyproject.toml
[tool.poetry]
name = "digstsgql"
version = "0.1.0"
description = "Semantic GraphQL project for Digitaliseringsstyrelsen"
authors = ["Magenta ApS <info@magenta.dk>"]

[tool.poetry.dependencies]
python = "^3.11"
more-itertools = "^10"
pydantic = "^2"
pydantic-settings = "^2"
strawberry-graphql = "^0.253"
starlette = "^0.41"
uvicorn = "^0.29"
sqlalchemy = "^2"
psycopg = "^3"

[tool.poetry.group.pre-commit.dependencies]
mypy = "^1"
pre-commit = "^3"
ruff = "^0.7"
deptry = "^0.16"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["pydantic.mypy", "strawberry.ext.mypy_plugin"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.deptry.per_rule_ignores]
DEP002 = [
  # uvicorn is used by the Dockerfile
  "uvicorn",
  # psycopg is used by sqlalchemy
  "psycopg",
]
